+++ ip_test.go
@@ -9,44 +9,73 @@ import (
 func TestParseForwardedHeader(t *testing.T) {
 	header := []string{
 		"for=12.34.56.78;host=example.com;proto=https, for=23.45.67.89",
+		"for=12.34.56.78, for=23.45.67.89;secret=egah2CGj55fSJFs, for=65.182.89.102",
 		"for=12.34.56.78, for=23.45.67.89;secret=egah2CGj55fSJFs, for=10.1.2.3",
 		"for=192.0.2.60;proto=http;by=203.0.113.43",
+		"for=10.1.2.3;proto=http;by=203.0.113.43",
 		"proto=http;by=203.0.113.43;for=192.0.2.61",
+		"proto=http;by=203.0.113.43;for=10.1.2.3",
 		"   ",
 		"",
 	}
 	expected := []string{
-		"12.34.56.78",
-		"12.34.56.78",
+		"23.45.67.89",
+		"65.182.89.102",
+		"",
 		"192.0.2.60",
+		"",
 		"192.0.2.61",
+		"",
 		"",
 		"",
 	}
 
 	for i, head := range header {
-		assert.Equal(t, expected[i], parseForwardedHeader(head))
+		assert.Equal(t, expected[i], ParseForwardedHeader(head))
 	}
 }
 
 func TestParseXForwardedForHeader(t *testing.T) {
 	header := []string{
-		"127.0.0.1",
-		"127.0.0.1, 23.21.45.67",
-		"127.0.0.1,23.21.45.67",
+		"65.182.89.102",
+		"127.0.0.1, 23.21.45.67, 65.182.89.102",
+		"127.0.0.1,23.21.45.67,65.182.89.102",
+		"65.182.89.102,23.21.45.67,127.0.0.1",
 		"   ",
 		"",
 	}
 	expected := []string{
-		"127.0.0.1",
-		"127.0.0.1",
-		"127.0.0.1",
+		"65.182.89.102",
+		"65.182.89.102",
+		"65.182.89.102",
+		"",
 		"",
 		"",
 	}
 
 	for i, head := range header {
-		assert.Equal(t, expected[i], parseXForwardedForHeader(head))
+		assert.Equal(t, expected[i], ParseXForwardedForHeader(head))
+	}
+}
+
+func TestParseXRealIPHeader(t *testing.T) {
+	header := []string{
+		"",
+		"  ",
+		"invalid",
+		"127.0.0.1",
+		"65.182.89.102",
+	}
+	expected := []string{
+		"",
+		"",
+		"",
+		"",
+		"65.182.89.102",
+	}
+
+	for i, head := range header {
+		assert.Equal(t, expected[i], ParseXRealIPHeader(head))
 	}
 }
 
@@ -66,14 +95,14 @@ func TestGetIP(t *testing.T) {
 	assert.Equal(t, "192.0.2.60", getIP(r))
 
 	// X-Forwarded-For
-	r.Header.Set("X-Forwarded-For", "127.0.0.1, 23.21.45.67")
-	assert.Equal(t, "127.0.0.1", getIP(r))
+	r.Header.Set("X-Forwarded-For", "127.0.0.1, 23.21.45.67, 65.182.89.102")
+	assert.Equal(t, "65.182.89.102", getIP(r))
 
 	// True-Client-IP
-	r.Header.Set("True-Client-IP", "127.0.0.1, 23.21.45.67")
-	assert.Equal(t, "127.0.0.1", getIP(r))
+	r.Header.Set("True-Client-IP", "127.0.0.1, 23.21.45.67, 65.182.89.102")
+	assert.Equal(t, "65.182.89.102", getIP(r))
 
 	// CF-Connecting-IP
-	r.Header.Set("CF-Connecting-IP", "127.0.0.1, 23.21.45.67")
-	assert.Equal(t, "127.0.0.1", getIP(r))
+	r.Header.Set("CF-Connecting-IP", "127.0.0.1, 23.21.45.67, 65.182.89.102")
+	assert.Equal(t, "65.182.89.102", getIP(r))
 }